+++ pytorch_lightning/trainer/connectors/logger_connector/result.py
@@ -19,7 +19,7 @@ from typing import Any, Callable, Dict, Iterable, List, Mapping, Optional, Tuple
 import torch
 from torchmetrics import Metric
 
-from pytorch_lightning.core.mixin import DeviceDtypeModuleMixin
+from pytorch_lightning.core.mixins import DeviceDtypeModuleMixin
 from pytorch_lightning.utilities import rank_zero_warn
 from pytorch_lightning.utilities.apply_func import apply_to_collection, apply_to_collections
 from pytorch_lightning.utilities.distributed import distributed_available
