+++ tests/utilities/test_dtype_device_mixin.py
@@ -16,7 +16,7 @@ import torch
 import torch.nn as nn
 
 from pytorch_lightning import Callback, Trainer
-from pytorch_lightning.core.mixin import DeviceDtypeModuleMixin
+from pytorch_lightning.core.mixins import DeviceDtypeModuleMixin
 from tests.helpers import BoringModel
 from tests.helpers.runif import RunIf
